ml_training/requirements.txt
requirements.txt file for ML training

Dependencies for model training with PyTorch:

torch==2.1.1
torchvision==0.16.1
pytorch-lightning==2.1.1
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
opencv-python==4.8.1.78
jupyter==1.0.0
